requirements.txt
requests
anthropic
python-dotenv
reportlab
shapely
